+++ internal/metric/domain/model.go
@@ -1,4 +1,4 @@
-package metric
+package domain
 
 import "time"
